allow command + a to be accessible also on table and board focus
