+++ packages/twenty-front/src/modules/object-record/record-board/components/RecordBoardHotkeyEffect.tsx
@@ -3,10 +3,15 @@ import { Key } from 'ts-key-enum';
 
 import { RecordBoardContext } from '@/object-record/record-board/contexts/RecordBoardContext';
 import { useBoardCardNavigation } from '@/object-record/record-board/hooks/useBoardCardNavigation';
+import { useRecordBoardSelection } from '@/object-record/record-board/hooks/useRecordBoardSelection';
 import { BoardHotkeyScope } from '@/object-record/record-board/types/BoardHotkeyScope';
+import { recordIndexAllRecordIdsComponentSelector } from '@/object-record/record-index/states/selectors/recordIndexAllRecordIdsComponentSelector';
 import { RecordIndexHotkeyScope } from '@/object-record/record-index/types/RecordIndexHotkeyScope';
 import { usePreviousHotkeyScope } from '@/ui/utilities/hotkey/hooks/usePreviousHotkeyScope';
 import { useScopedHotkeys } from '@/ui/utilities/hotkey/hooks/useScopedHotkeys';
+import { getSnapshotValue } from '@/ui/utilities/recoil-scope/utils/getSnapshotValue';
+import { useRecoilComponentCallbackStateV2 } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentCallbackStateV2';
+import { useRecoilCallback } from 'recoil';
 
 export const RecordBoardHotkeyEffect = () => {
   const { recordBoardId } = useContext(RecordBoardContext);
@@ -14,6 +19,35 @@ export const RecordBoardHotkeyEffect = () => {
   const { move } = useBoardCardNavigation(recordBoardId);
 
   const { setHotkeyScopeAndMemorizePreviousScope } = usePreviousHotkeyScope();
+
+  const recordIndexAllRecordIdsState = useRecoilComponentCallbackStateV2(
+    recordIndexAllRecordIdsComponentSelector,
+  );
+
+  const { setRecordAsSelected } = useRecordBoardSelection(recordBoardId);
+
+  const selectAll = useRecoilCallback(
+    ({ snapshot }) =>
+      () => {
+        const allRecordIds = getSnapshotValue(
+          snapshot,
+          recordIndexAllRecordIdsState,
+        );
+
+        for (const recordId of allRecordIds) {
+          setRecordAsSelected(recordId, true);
+        }
+      },
+    [recordIndexAllRecordIdsState, setRecordAsSelected],
+  );
+
+  useScopedHotkeys(
+    'ctrl+a,meta+a',
+    selectAll,
+    RecordIndexHotkeyScope.RecordIndex,
+  );
+
+  useScopedHotkeys('ctrl+a,meta+a', selectAll, BoardHotkeyScope.BoardFocus);
 
   useScopedHotkeys(
     Key.ArrowLeft,
